requirements.txt
pyaudio
numpy
python-dotenv
websocket-client
customtkinter
